Init - Project Structure Fixes
