+++ data_processor.py
@@ -1,0 +1,39 @@
+import pandas as pd
+import numpy as np
+from sklearn.preprocessing import MinMaxScaler
+from sklearn.model_selection import train_test_split
+
+def preprocess_data(df):
+    df['SMA_50'] = df['Close'].rolling(window=50).mean()
+    df['EMA_50'] = df['Close'].ewm(span=20, adjust=False).mean()
+    df['Daily_Return'] = df['Close'].pct_change()
+    df['Log_Return'] = np.log(df['Close'] / df['Close'].shift(1))
+
+    delta = df['Close'].diff(1)
+    gain = (delta.where(delta > 0, 0)).rolling(window=14).mean()
+    loss = (-delta.where(delta < 0, 0)).rolling(window=14).mean()
+    rs = gain / loss
+
+    df['RSI_14'] = 100 - (100 / (1 + rs))
+    df['Middle_Band'] = df['Close'].rolling(window=20).mean()
+    df['Upper_Band'] = df['Middle_Band'] + (df['Close'].rolling(window=20).std() * 2)
+    df['Lower_Band'] = df['Middle_Band'] - (df['Close'].rolling(window=20).std() * 2)
+    df['Momemtum'] = df['Close'] - df['Close'].shift(4)
+    df['Volatility'] = df['Close'].rolling(window=21).std()
+
+    df.dropna(inplace=True)
+    return df
+
+def normalize_data(df):
+    scaler = MinMaxScaler()
+    scaled_data = scaler.fit_transform(df[['Close']])
+    return scaled_data, scaler
+
+def prepare_data(scaled_data, time_steps=60):
+    X, y = [], []
+    for i in range(time_steps, len(scaled_data)):
+        X.append(scaled_data[i - time_steps: i])
+        y.append(scaled_data[i, 0])
+
+    print(f"X Array Shape: {np.array(X).shape}, Y Array Shape: {np.array(y).shape}")
+    return np.array(X), np.array(y)
