+++ model_builder.py
@@ -1,0 +1,12 @@
+from keras.models import Sequential
+from keras.layers import LSTM, Dense
+
+def build_model(input_shape):
+    model = Sequential([
+        LSTM(50, return_sequences=True, input_shape=input_shape),
+        LSTM(50, return_sequences=False),
+        Dense(25),
+        Dense(1)
+    ])
+    model.compile(optimizer='adam', loss='mean_squared_error')
+    return model
